merge
Former-commit-id: 97e23f7f39ac407a689799b3187b74d091d904f9

+++ client/app/components/Admin/AdminPricingRouteView.jsx
@@ -170,8 +170,9 @@ export class AdminPricingRouteView extends Component {
             <div
               className="flex-5 layout-row layout-align-center-center"
               onClick={() => {
-                console.log('delete')
-                adminActions.deletePricing(pricing)
+                if (window.confirm('Are you sure you want to delete this pricing?')) {
+                  adminActions.deletePricing(pricing)
+                }
               }}
             >
               <i className="flex-none fa fa-trash pointy" />
